create and update lessons

+++ E-learning/src/main/java/com/elearning/Sections/SectionService.java
@@ -1,5 +1,7 @@
 package com.elearning.Sections;
 
+import com.elearning.Courses.Course;
+import com.elearning.Courses.CourseAuthorization;
 import com.elearning.Courses.Repo.CourseJpaRepo;
 import com.elearning.Courses.CourseService;
 import com.elearning.Exceptions.BadRequestException;
@@ -7,6 +9,7 @@ import com.elearning.Exceptions.NotAllowedOperation;
 import com.elearning.Exceptions.NotFoundException;
 import com.elearning.Exceptions.UnAuthorizedException;
 import com.elearning.entities.Section;
+import com.elearning.entities.users.User;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -19,14 +22,25 @@ public class SectionService {
     private final CourseService courseService;
     private final SectionJpaRepo sectionJpaRepo;
     private final CourseJpaRepo courseJpaRepo;
+    private final CourseAuthorization courseAuthorization;
+
+    private void checkWriteAccess(Long courseId,Long usrId)
+    {
+        var course = courseService.findById(courseId);
+        User currUsr = new User();
+        currUsr.setId(usrId);
+        checkWriteAccess(currUsr,course);
+    }
+    private void checkWriteAccess(User currUsr, Course course)
+    {
+        // handle authorization better way later
+        if(courseAuthorization.canWrite(currUsr,course)){
+            throw new UnAuthorizedException("Not authorized to edit this course");
+        }
+    }
     private void courseExists(Long courseId){
         if(!courseService.isExists(courseId)){
             throw new NotFoundException("Course with id:" + courseId +" is not found");
-        }
-    }
-    private void canEditCourse(Long usrId , Long courseId){
-        if(!courseService.canEditCourse(usrId,courseId)){
-            throw new UnAuthorizedException("Un authorized to edit this course");
         }
     }
     private void checkSectionIndex(Long courseId , Integer index)
@@ -34,15 +48,19 @@ public class SectionService {
         if(existsByCourseIdAndSectionIdx(courseId,index))
             throw new BadRequestException("section index already exists");
     }
+    public  Section findById(Long sectionId)
+    {
+        return sectionJpaRepo.findById(sectionId).orElseThrow(() -> new NotFoundException("section with id:" + sectionId +" does not exist"));
+    }
     public Collection<Section> getCourseSections(Long courseId)
     {
         courseExists(courseId);
         return sectionJpaRepo.findAllByCourse_Id(courseId);
     }
-    public Section createSection(Long usrId , Long courseId,Integer index , String title){
-        courseExists(courseId);
-        // handle authorization better way later
-        canEditCourse(usrId,courseId);
+    public Section createSection(Long courseId,Integer index , String title){
+
+        checkWriteAccess(courseId,1L);
+
         checkSectionIndex(courseId,index);
         Section section = new Section();
         section.setCourse(courseJpaRepo.getReferenceById(courseId));
@@ -57,9 +75,12 @@ public class SectionService {
         return sectionJpaRepo.existsByCourse_IdAndIndex(courseId,index);
     }
 
-    public Section updateSection(Long usrId ,Long courseId,Long sectionId,Integer index , String title){
-        courseExists(courseId);
-        canEditCourse(usrId,courseId);
+    public boolean isCourseAndSectionExists(Long courseId,Long sectionId)
+    {
+        return sectionJpaRepo.existsByIdAndCourse_Id(sectionId,courseId);
+    }
+    public Section updateSection(Long courseId,Long sectionId,Integer index , String title){
+        checkWriteAccess(courseId,1L);
         checkSectionIndex(courseId,index);
         var section = sectionJpaRepo.findById(sectionId).orElseThrow(() -> new NotFoundException("Section with id:" + sectionId + " does not exists"));
         section.setTitle(title);
@@ -73,9 +94,8 @@ public class SectionService {
     {
         return isEmptySection(sectionId);
     }
-    public void deleteSection(Long usrId , Long courseId , Long sectionId){
-
-        canEditCourse(usrId,courseId);
+    public void deleteSection(Long courseId , Long sectionId){
+        checkWriteAccess(courseId,1L);
         if(isSectionDeletable(sectionId))
             sectionJpaRepo.deleteById(sectionId);
         else
